Crate syn_solidity

syn-solidity

syn-powered parser for Solidity-like TokenStreams.

The parsed root element is the File, which contains a list of Items. Items also support outer attributes, as shown below.

⚠️ Work in progress ⚠️

Design

This parser is specifically designed for Rust procedural macros. It aims to mimic the behavior of the official Solidity compiler (Solc) when it comes to parsing valid Solidity code. This means that all valid Solidity code, as recognized by Solc v0.5.*1 and above, will also be recognized and parsed correctly by syn-solidity.

However, syn-solidity is more permissive and lenient compared to the official Solidity compiler and grammar specifications. Some examples of code patterns that are valid in syn-solidity but not in the official compiler include:

identifiers are Rust identifiers (syn::Ident), and as such cannot contain the dollar sign ($), but can contain unicode characters

trailing punctuation, like commas (,) in function arguments or enums definitions

certain variable and function attributes in certain contexts, like internal functions or functions with implementations ({ ... }) in interfaces

parameter storage locations in item definitions, like uint256[] memory in a struct or error definition

the tuple type (T, U, ..) is allowed wherever a type is expected, and can optionally be preceded by the tuple keyword. This is the same as ethers.js’s Human-Readable ABI

This lenient behavior is intentionally designed to facilitate usage within procedural macros, and to reduce general code complexity in the parser and AST.

Example

use quote::quote;
use syn_solidity::{File, Item};

// Create a Solidity `TokenStream`
let tokens = quote! {
    /// @name HelloWorld
    /// @notice A hello world example in Solidity.
    contract HelloWorld {
        /// @notice Returns the string "Hello, World!".
        function helloWorld() external pure returns (string memory) {
            return "Hello, World!";
        }
    }
};

// Parse the tokens into a `File`
let ast: File = syn_solidity::parse2(tokens)?;

let items: &[Item] = &ast.items;
let Some(Item::Contract(contract)) = items.first() else { unreachable!() };
assert_eq!(contract.name, "HelloWorld");
assert_eq!(contract.attrs.len(), 2);

let body: &[Item] = &contract.body;
let Some(Item::Function(function)) = items.first() else { unreachable!() };
assert_eq!(function.attrs.len(), 1);
assert_eq!(function.name, "helloWorld");
assert!(function.arguments.is_empty());
assert_eq!(function.attributes.len(), 3);

Older versions may still parse successfully, but this is not guaranteed. ↩

Re-exports

pub use visit::Visit;

pub use visit_mut::VisitMut;

Modules

kw

Solidity keywords.

visit

Syntax tree traversal to walk a shared borrow of a syntax tree.

visit_mut

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place.

Macros

sol_path

Create a SolPath from a list of identifiers.

Structs

Block

A curly-braced block of statements.

EventParameter

An event parameter.

File

A Solidity file. The root of the AST.

FunctionAttributes

A list of unique function attributes. Used in ItemFunction.

ImportAlias

An import alias.

ImportAliases

A list of import aliases: import { Foo as Bar, Baz } from "foo.sol";

ImportDirective

An import directive: import "foo.sol";

ImportGlob

A glob import directive: import * as Foo from "foo.sol";

ImportPlain

A plain import directive: import "foo.sol" as Foo;

ItemContract

A contract, abstract contract, interface, or library definition: contract Foo is Bar("foo"), Baz { ... }

ItemEnum

An enum definition: enum Foo { A, B, C }

ItemError

An error definition: error Foo(uint256 a, uint256 b);

ItemEvent

ItemFunction

A function, constructor, fallback, receive, or modifier definition: function helloWorld() external pure returns(string memory);

ItemStruct

A struct definition: struct Foo { uint256 bar; }

ItemUdt

A user-defined value type definition: type Foo is uint256;

LitStr

A string literal.

Modifier

A modifier invocation, or an inheritance specifier.

Override

The override attribute.

Parameters

A list of VariableDeclarations, separated by P.

PragmaDirective

A pragma directive: pragma solidity ^0.8.0;

Returns

The returns attribute of a function.

SolIdent

A Solidity identifier.

SolPath

A list of identifiers, separated by dots.

TypeArray

An array type.

TypeFunction

A function type: function() returns (string memory)

TypeMapping

A mapping type: mapping(uint key => string value)

TypeTuple

A tuple type.

UsingDirective

A using directive: using { A, B.mul as * } for uint256 global;

UsingListItem

VariableAttributes

A list of unique variable attributes.

VariableDeclaration

A variable declaration.

VariableDefinition

Enums

FunctionAttribute

A function attribute.

ImportPath

The path of an import directive.

Item

An AST item. A more expanded version of a Solidity source unit.

Mutability

A mutability attribute.

PragmaTokens

Storage

A storage location.

Type

A type name.

UserDefinableOperator

A user-definable operator: +, *, |, etc.

UsingList

UsingType

VariableAttribute

A variable attribute.

Visibility

A visibility attribute.

Functions

parse

Parse a Solidity proc_macro::TokenStream into a File.

parse2

Parse a Solidity proc_macro2::TokenStream into a File.

Type Definitions

FieldList

A list of semicolon-separated VariableDeclarations.

ParameterList

A list of comma-separated VariableDeclarations.