syn_solidity

Struct ItemEvent

pub struct ItemEvent {
    pub attrs: Vec<Attribute>,
    pub event_token: event,
    pub name: SolIdent,
    pub paren_token: Paren,
    pub parameters: Punctuated<EventParameter, Comma>,
    pub anonymous: Option<anonymous>,
    pub semi_token: Semi,
}

Fields
attrs: Vec<Attribute>event_token: eventname: SolIdentparen_token: Parenparameters: Punctuated<EventParameter, Comma>anonymous: Option<anonymous>semi_token: Semi
Implementations

impl ItemEvent

pub const fn is_anonymous(&self) -> bool

Returns true if the event is anonymous.

pub fn max_indexed(&self) -> usize

Returns the maximum amount of indexed parameters this event can have.

This is 4 if the event is anonymous, otherwise 3.

pub fn exceeds_max_indexed(&self) -> bool

Returns true if the event has more indexed parameters than allowed by Solidity.

See Self::max_indexed.

pub fn assert_valid(&self) -> Result<()>

Asserts that the event has a valid amount of indexed parameters.

pub fn params(&self) -> ParameterList

pub fn param_types( &self, ) -> impl ExactSizeIterator<Item = &Type> + DoubleEndedIterator + Clone

pub fn param_types_mut( &mut self, ) -> impl ExactSizeIterator<Item = &mut Type> + DoubleEndedIterator

pub fn param_types_and_names( &self, ) -> impl ExactSizeIterator<Item = (&Type, Option<&SolIdent>)> + DoubleEndedIterator

pub fn param_type_strings( &self, ) -> impl ExactSizeIterator<Item = String> + DoubleEndedIterator + Clone + '_

pub fn non_indexed_params(&self) -> impl Iterator<Item = &EventParameter>

pub fn indexed_params(&self) -> impl Iterator<Item = &EventParameter>

pub fn as_type(&self) -> Type