syn_solidity

Struct ItemFunction

pub struct ItemFunction {
    pub attrs: Vec<Attribute>,
    pub kind: FunctionKind,
    pub name: Option<SolIdent>,
    pub paren_token: Option<Paren>,
    pub parameters: ParameterList,
    pub attributes: FunctionAttributes,
    pub returns: Option<Returns>,
    pub body: FunctionBody,
}

A function, constructor, fallback, receive, or modifier definition: function helloWorld() external pure returns(string memory);.

Solidity reference: https://docs.soliditylang.org/en/latest/grammar.html#a4.SolidityParser.functionDefinition

Fields
attrs: Vec<Attribute>

The syn attributes of the function.

kind: FunctionKindname: Option<SolIdent>paren_token: Option<Paren>

Parens are optional for modifiers: https://docs.soliditylang.org/en/latest/grammar.html#a4.SolidityParser.modifierDefinition

parameters: ParameterListattributes: FunctionAttributes

The Solidity attributes of the function.

returns: Option<Returns>

The optional return types of the function.

body: FunctionBody
Implementations

impl ItemFunction

pub fn new(kind: FunctionKind, name: Option<SolIdent>) -> Self

Create a new function of the given kind.

pub fn new_getter(name: SolIdent, ty: Type) -> Self

Create a new function with the given name and arguments.

Note that:

the type is not validated

structs/array of structs in return position are not expanded

the body is not set

The attributes are set to public view.

See the Solidity documentation for more details on how getters are generated.

pub fn from_variable_definition(var: VariableDefinition) -> Self

Creates a new function from a variable definition.

The function will have the same name and the variable typeโs will be the return type. The variable attributes are ignored, and instead will always generate public returns.

See new_getter for more details.

pub fn name(&self) -> &SolIdent

Returns the name of the function.

Panics

Panics if the function has no name. This is the case when kind is not Function.

pub fn is_void(&self) -> bool

Returns true if the function returns nothing.

pub fn has_implementation(&self) -> bool

Returns true if the function has a body.

pub fn call_type(&self) -> Type

Returns the functionโs arguments tuple type.

pub fn return_type(&self) -> Option<Type>

Returns the functionโs return tuple type.

pub fn body(&self) -> Option<&[Stmt]>

Returns a reference to the functionโs body, if any.

pub fn body_mut(&mut self) -> Option<&mut Vec<Stmt>>

Returns a mutable reference to the functionโs body, if any.

pub fn into_body(self) -> Result<Vec<Stmt>, Self>