Module syn_solidity::visit_mut

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place.

Each method of the VisitMut trait is a hook that can be overridden to customize the behavior when visiting the corresponding type of node. By default, every method recursively visits the substructure of the input by invoking the right visitor method of each of its fields.

Traits

VisitMut

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place.

Functions

visit_field_list

visit_file

visit_ident

visit_import_directive

visit_item

visit_item_contract

visit_item_enum

visit_item_error

visit_item_event

visit_item_function

visit_item_struct

visit_item_udt

visit_lit_str

visit_parameter_list

visit_path

visit_pragma_directive

visit_type

visit_using_directive

visit_variable_declaration

visit_variable_definition