syn_solidity

Module visit_mut

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place.

Each method of the VisitMut trait is a hook that can be overridden to customize the behavior when visiting the corresponding type of node. By default, every method recursively visits the substructure of the input by invoking the right visitor method of each of its fields.

Traits

VisitMut

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place.

Functions

visit_block

visit_expr

visit_expr_array

visit_expr_binary

visit_expr_call

visit_expr_call_options

visit_expr_delete

visit_expr_index

visit_expr_member

visit_expr_new

visit_expr_payable

visit_expr_postfix

visit_expr_ternary

visit_expr_tuple

visit_expr_type_call

visit_expr_unary

visit_field_list

visit_file

visit_ident

visit_import_directive

visit_item

visit_item_contract

visit_item_enum

visit_item_error

visit_item_event

visit_item_function

visit_item_struct

visit_item_udt

visit_lit

visit_lit_denominated

visit_lit_str

visit_parameter_list

visit_path

visit_pragma_directive

visit_stmt

visit_stmt_asm

visit_stmt_break

visit_stmt_continue

visit_stmt_dowhile

visit_stmt_emit

visit_stmt_for

visit_stmt_if

visit_stmt_return

visit_stmt_revert

visit_stmt_try

visit_stmt_var_decl

visit_stmt_while

visit_type

visit_unchecked_block

visit_using_directive

visit_variable_declaration

visit_variable_definition